Compute bootstrap confidence intervals on the average squared density parameter

Compute bootstrap confidence intervals on the average squared density parameter

Compute bootstrap confidence intervals on the average squared density parameter

Details

Output both the point estimate and the bootstrap confidence interval

Super class

TMLEbootstrap::generalBootstrap -> avgDensityBootstrap

Methods

Public methods

avgDensityBootstrap$new()

avgDensityBootstrap$bootstrap_once()

avgDensityBootstrap$run_bootstrap()

avgDensityBootstrap$clone()

Inherited methods

TMLEbootstrap::generalBootstrap$all_boot_CI()

TMLEbootstrap::generalBootstrap$bootstrap()

TMLEbootstrap::generalBootstrap$center_CI()

TMLEbootstrap::generalBootstrap$exact_bootstrap()

TMLEbootstrap::generalBootstrap$exact_bootstrap_paper()

TMLEbootstrap::generalBootstrap$penalized_CI()

TMLEbootstrap::generalBootstrap$penalized_CI_half()

TMLEbootstrap::generalBootstrap$scale_adjust_CI()

TMLEbootstrap::generalBootstrap$shift2()

TMLEbootstrap::generalBootstrap$sigma_mse()

TMLEbootstrap::generalBootstrap$spread()

Method new()

bootstrap average density parameter;

Usage

avgDensityBootstrap$new(
  x,
  epsilon_step = NULL,
  bin_width = 0.3,
  lambda_grid = NULL,
  M = NULL,
  targeting = TRUE
)

Arguments

x

vector of random samples.

epsilon_step

step size of the TMLE (one-step)

bin_width

binning width for the density super learner

lambda_grid

pre-specified grid of L1 penalization for the HAL

M

alternative to "lambda_grid". Specify the max variation norm of the HAL

targeting

bool. whether to perform targeting on the density estimator. If FALSE, the density super learner estimate will be returned

Returns

NULL

Method bootstrap_once()

run one bootstrap sample

Usage

avgDensityBootstrap$bootstrap_once(
  self,
  data,
  epsilon_step = self$epsilon_step,
  population_x = self$x,
  population_tmle = self$pointTMLE,
  inflate_lambda = 1
)

Arguments

self

self

data

input bootstrap data

epsilon_step

step size of the TMLE (one-step)

population_x

whole training data

population_tmle

TMLE on the whole training data

inflate_lambda

experimental. scalar to scale the L1 penalty during the bootstrap

Returns

NULL

Method run_bootstrap()

Usage

avgDensityBootstrap$run_bootstrap(
  n_bootstrap = 200,
  alpha = 0.05,
  kind = NULL,
  inflate_lambda = 1,
  to_parallel = FALSE
)

Method clone()

The objects of this class are cloneable with this method.

Usage

avgDensityBootstrap$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.