Compute TMLE on the variance of CATE (continuous Y)

inherit the initial fit routine from ateTMLE replace initial fit with continuous hal9001; rescales Y into (0,1) before TMLE; scales the Psi, EIC, CI back to the original scale after the TMLE scale Y to (0,1)

Super classes

TMLEbootstrap::ateTMLE -> TMLEbootstrap::blipVarianceTMLE -> blipVarianceTMLEContinuousY

Methods

Public methods

blipVarianceTMLEContinuousY$scale_outcome()

blipVarianceTMLEContinuousY$scale_back_after_tmle()

blipVarianceTMLEContinuousY$initial_fit()

blipVarianceTMLEContinuousY$target()

blipVarianceTMLEContinuousY$inference_without_target()

blipVarianceTMLEContinuousY$clone()

Inherited methods

TMLEbootstrap::ateTMLE$initial_fit_pen_likeli()

TMLEbootstrap::ateTMLE$initialize()

TMLEbootstrap::ateTMLE$plot_Q1W()

TMLEbootstrap::blipVarianceTMLE$compute_EIC()

TMLEbootstrap::blipVarianceTMLE$compute_min_phi_ratio()

Method scale_outcome()

Usage

blipVarianceTMLEContinuousY$scale_outcome()

Method scale_back_after_tmle()

Usage

blipVarianceTMLEContinuousY$scale_back_after_tmle()

Method initial_fit()

Usage

blipVarianceTMLEContinuousY$initial_fit(...)

Method target()

Usage

blipVarianceTMLEContinuousY$target()

Method inference_without_target()

Usage

blipVarianceTMLEContinuousY$inference_without_target()

Method clone()

The objects of this class are cloneable with this method.

Usage

blipVarianceTMLEContinuousY$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.