Abstract class of bootstrap

Abstract class of bootstrap

Abstract class of bootstrap

Details

Output both the point estimate and the bootstrap confidence interval

Methods

Public methods

generalBootstrap$new()

generalBootstrap$bootstrap()

generalBootstrap$exact_bootstrap()

generalBootstrap$exact_bootstrap_paper()

generalBootstrap$center_CI()

generalBootstrap$scale_adjust_CI()

generalBootstrap$penalized_CI()

generalBootstrap$penalized_CI_half()

generalBootstrap$shift2()

generalBootstrap$sigma_mse()

generalBootstrap$spread()

generalBootstrap$all_boot_CI()

generalBootstrap$clone()

Method new()

Usage

generalBootstrap$new()

Method bootstrap()

regular bootstrap

Usage

generalBootstrap$bootstrap(n_bootstrap, alpha = 0.05, ...)

Arguments

n_bootstrap

number of bootstrap samples

alpha

significance level for CI

...

other passed to "run_bootstrap"

Returns

NULL

Method exact_bootstrap()

exact bootstrap

Usage

generalBootstrap$exact_bootstrap(n_bootstrap, alpha = 0.05, ...)

Arguments

n_bootstrap

number of bootstrap samples

alpha

significance level for CI

...

other passed to "run_bootstrap"

Returns

NULL

Method exact_bootstrap_paper()

exact bootstrap on vdl paper

Usage

generalBootstrap$exact_bootstrap_paper(n_bootstrap, alpha = 0.05, ...)

Arguments

n_bootstrap

number of bootstrap samples

alpha

significance level for CI

...

other passed to "run_bootstrap"

Returns

NULL

Method center_CI()

center the bootstrap CI on the point estimate

Usage

generalBootstrap$center_CI(bootCI = NULL)

Arguments

bootCI

the vector of CI

Returns

NULL

Method scale_adjust_CI()

scale the bootstrap CI to be wider than Wald CI

Usage

generalBootstrap$scale_adjust_CI(bootCI = NULL)

Arguments

bootCI

the vector of CI

Returns

NULL

Method penalized_CI()

bias penalized bootstrap

add abs(bias) to both sides of the bootstrap CI (wider)

Usage

generalBootstrap$penalized_CI(bootCI = NULL)

Arguments

bootCI

the vector of CI

Returns

NULL

Method penalized_CI_half()

bias penalized bootstrap (half)

add 0.5*abs(bias) to both sides of the bootstrap CI (wider)

Usage

generalBootstrap$penalized_CI_half(bootCI = NULL)

Arguments

bootCI

the vector of CI

Returns

NULL

Method shift2()

two bias shifted bootstrap

shift the center of CI by two times the bias

Usage

generalBootstrap$shift2(bootCI = NULL)

Arguments

bootCI

the vector of CI

Returns

NULL

Method sigma_mse()

use the stddev of bootstrap samples, to put a Normal CI around point estimate

Usage

generalBootstrap$sigma_mse(bootCI = NULL)

Arguments

bootCI

the vector of CI

Returns

CI

Method spread()

use the z-quantiles of bootstrap samples, to put a Normal CI around point estimate

Usage

generalBootstrap$spread(bootCI = NULL)

Arguments

bootCI

the vector of CI

Returns

CI

Method all_boot_CI()

return all CI

Usage

generalBootstrap$all_boot_CI()

Returns

list of CI

Method clone()

The objects of this class are cloneable with this method.

Usage

generalBootstrap$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.